Separate out viewlist
