+++ jupyter-widgets-base/src/index.ts
@@ -7,3 +7,4 @@ export * from "./widget";
 export * from "./widget_layout";
 export * from "./widget_style";
 export * from "./services-shim";
+export * from "./viewlist";
